+++ src/index.js
@@ -1,7 +1,7 @@
 import { ApplyRove } from './ApplyRoving';
 
 export const RovueingTabindex = {
-    install(app, { direction }) {
+    install(app, { direction = 'ltr' } = {}) {
         const roveMap = new WeakMap();
         app.directive('rove', {
             created(el, { value = true, modifiers: { rtl = direction === 'rtl' } }) {
@@ -16,9 +16,15 @@ export const RovueingTabindex = {
                 const observerConfig = { attributes: false, childList: true };
                 roveMap.set(el, { observer, observerConfig });
             },
-            mounted(el) {
+            mounted(el, { value = true, modifiers: { rtl = direction === 'rtl' } }) {
                 const { observer, observerConfig } = roveMap.get(el);
                 observer.observe(el, observerConfig);
+                ApplyRove({
+                    element: el,
+                    disabled: !value,
+                    isRTL: rtl,
+                    roveMap
+                });
             },
             updated(el) {
                 const { observer, observerConfig } = roveMap.get(el);