+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [UNRELEASED]
+### Fixed
+- Fix missing `direction`
+
 ## [0.0.1] - 2021-05-05
 ### Added
 - Initial release